requirements.txt
# base ----------------------------------------
flask
gunicorn
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless>=4.1.2
Pillow
psutil
PyYAML>=5.3.1
scipy>=1.4.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.12.0
tqdm>=4.41.0

# logging -------------------------------------
##tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.9.0
# scikit-learn==0.19.2  # for coreml quantization
# tensorflow==2.4.1  # for TFLite export

# extras --------------------------------------
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
thop  # FLOPS computation
IPython